+++ test/001-default.ini
@@ -1,4 +1,4 @@
-# This file should remain empty, except for the motherbooard.
+# This file should remain empty except for the motherboard.
 # If changes are needed by tests, it should be performed in another configuration.
 
 [config:base]
