+++ test/002-extruders_1_runout.ini
@@ -9,10 +9,10 @@ motherboard                = BOARD_SIMULATED
 
 # Options to support runout sensors test
 filament_runout_sensor     = on
-fil_runout_pin             = 4
+fil_runout_pin             = 4  # dummy
 advanced_pause_feature     = on
 emergency_parser           = on
 nozzle_park_feature        = on
 
-# Option to support testing parsing with paranthesis enabled in comments
+# Option to support testing parsing with parentheses comments enabled
 paren_comments             = on
