+++ test/003-extruders_3_runout.ini
@@ -1,0 +1,32 @@
+#
+# Test configuration with three extruders and filament runout sensors
+#
+[config:base]
+ini_use_config             = base
+
+# Unit tests must use BOARD_SIMULATED to run natively in Linux
+motherboard                = BOARD_SIMULATED
+
+# Options to support runout sensor tests on three extruders.
+# Options marked "dummy" are simply required to pass sanity checks.
+extruders                  = 3
+temp_sensor_1              = 1
+temp_sensor_2              = 1
+temp_2_pin                 = 4 # dummy
+temp_3_pin                 = 4 # dummy
+heater_2_pin               = 4 # dummy
+e2_step_pin                = 4 # dummy
+e2_dir_pin                 = 4 # dummy
+e2_enable_pin              = 4 # dummy
+e3_step_pin                = 4 # dummy
+e3_dir_pin                 = 4 # dummy
+e3_enable_pin              = 4 # dummy
+num_runout_sensors         = 3
+filament_runout_sensor     = on
+fil_runout_pin             = 4 # dummy
+fil_runout2_pin            = 4 # dummy
+fil_runout3_pin            = 4 # dummy
+filament_runout_script     = "M600 %%c"
+advanced_pause_feature     = on
+emergency_parser           = on
+nozzle_park_feature        = on
