--- test/test_3_extruders_with_runout_sensors/config.ini
@@ -1,33 +1,0 @@
-#
-# Test configuration with three extruders and filament runout sensors
-#
-[config:base]
-ini_use_config             = base
-
-# Unit tests must use BOARD_SIMULATED to run natively in Linux
-motherboard                = BOARD_SIMULATED
-
-# Options to support runout sensor tests on three extruders
-# Many of these options are clearly invalid, but have to be set
-# to something to pass sanity checks
-extruders                  = 3
-temp_sensor_1              = 1
-temp_sensor_2              = 1
-temp_2_pin                 = 4
-temp_3_pin                 = 4
-heater_2_pin               = 4
-e2_step_pin                = 4
-e2_dir_pin                 = 4
-e2_enable_pin              = 4
-e3_step_pin                = 4
-e3_dir_pin                 = 4
-e3_enable_pin              = 4
-num_runout_sensors         = 3
-filament_runout_sensor     = on
-fil_runout_pin             = 4
-fil_runout2_pin            = 4
-fil_runout3_pin            = 4
-filament_runout_script     = "M600 %%c"
-advanced_pause_feature     = on
-emergency_parser           = on
-nozzle_park_feature        = on
